Refactor HTTP gateway interface types
This PR separates the types used to implement the `http_request`
and `http_request_update` canister methods into their own module.

Additionally the `upgrade` flag is introduce in preparation for the fix
the /faq redirect (which will need to set that flag).
