+++ src/internet_identity/internet_identity.did
@@ -23,6 +23,7 @@ type HttpResponse = record {
     status_code: nat16;
     headers: vec HeaderField;
     body: blob;
+    upgrade : opt bool;
     streaming_strategy: opt StreamingStrategy;
 };
