+++ src/internet_identity/src/state.rs
@@ -8,6 +8,7 @@ use ic_cdk::{call, trap};
 use ic_certified_map::{Hash, RbTree};
 use ic_stable_structures::DefaultMemoryImpl;
 use internet_identity::signature_map::SignatureMap;
+use internet_identity_interface::http_gateway::HeaderField;
 use internet_identity_interface::*;
 use std::cell::{Cell, RefCell};
 use std::collections::HashMap;
